// schemas/contactForm.ts
import { z } from 'zod';
export const contactFormSchema = z.object({
  name: z.string().min(1, 'Name is required'),
  email: z.string().email('Invalid email address'),
  subject: z.string().optional(),
  message: z.string().min(1, 'Message is required'),
});

export type TContactFormData = z.infer<typeof contactFormSchema>;